Updates from CR comments
