+++ org-code-javabuilder/lib/src/main/resources/main_painter.json
@@ -1,11 +1,7 @@
 {
   "source": {
-    "ScannerSample.java": {
-      "text": "import org.code.neighborhood.Painter;\nimport org.code.neighborhood.World;\nimport java.util.Scanner;\n\npublic class ScannerSample {\n\tpublic static void main(String[] args) {\n\t\tWorld world = new World(4);\n\t\tWorld.setInstance(world);\n\t\tPainter painter = new Painter();\n\t\tpainter.move();\n\t\tHelperClass.printWelcomeMessage();\n\t\tScanner console = new Scanner(System.in);\n\t\tSystem.out.println(\"Enter your name: \");\n\t\tString name = console.next();\n\t\tSystem.out.println(\"Hello \" + name);\n\t}\n}\n",
-      "visible": true
-    },
-    "HelperClass.java": {
-      "text": "\npublic class HelperClass {\n\tpublic static void printWelcomeMessage() {\n\t\tSystem.out.println(\"Hello!\");\n\t}\n}\n",
+    "PainterSample.java": {
+      "text": "import org.code.neighborhood.Painter;\nimport org.code.neighborhood.World;\nimport java.util.Scanner;\n\npublic class PainterSample {\n\tpublic static void main(String[] args) {\n\t\tWorld world = new World(4);\n\t\tWorld.setInstance(world);\n\t\tPainter painter = new Painter();\n\t\tpainter.move();\n\t}\n}\n",
       "visible": true
     }
   },
