+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/Grid.java
@@ -1,17 +1,21 @@
 package org.code.neighborhood;
 
+import org.code.protocol.OutputAdapter;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 
 public class Grid {
-  private GridSquare[][] grid;
-  private int width;
-  private int height;
+  private final GridSquare[][] grid;
+  private final int width;
+  private final int height;
+  private final OutputAdapter outputAdapter;
 
-  protected Grid(GridSquare[][] squares) {
+  protected Grid(GridSquare[][] squares, OutputAdapter outputAdapter) {
     this.grid = squares;
     this.height = squares.length;
     this.width = squares[0].length;
+    this.outputAdapter = outputAdapter;
   }
 
   public void printGrid() {
@@ -38,6 +42,24 @@ public class Grid {
     } else {
       throw new NeighborhoodRuntimeException(ExceptionKeys.GET_SQUARE_FAILED);
     }
+  }
+
+  /**
+   * Hides all buckets from the screen
+   */
+  public void hideBuckets() {
+    this.outputAdapter.sendMessage(
+        new NeighborhoodSignalMessage(NeighborhoodSignalKey.HIDE_BUCKETS, new HashMap<>())
+    );
+  }
+
+  /**
+   * Displays all buckets on the screen
+   */
+  public void showBuckets() {
+    this.outputAdapter.sendMessage(
+        new NeighborhoodSignalMessage(NeighborhoodSignalKey.SHOW_BUCKETS, new HashMap<>())
+    );
   }
 
   protected int getSize() {
