+++ org-code-javabuilder/neighborhood/src/main/java/org/code/neighborhood/GridFactory.java
@@ -3,6 +3,8 @@ package org.code.neighborhood;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+
+import org.code.protocol.OutputAdapter;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -12,6 +14,11 @@ public class GridFactory {
   private static final String GRID_SQUARE_TYPE_FIELD = "tileType";
   private static final String GRID_SQUARE_ASSET_ID_FIELD = "assetId";
   private static final String GRID_SQUARE_VALUE_FIELD = "value";
+  private final OutputAdapter outputAdapter;
+
+  protected GridFactory(OutputAdapter outputAdapter) {
+    this.outputAdapter = outputAdapter;
+  }
 
   protected Grid createGridFromJSON(String filename) throws IOException {
     File file = new File(GRID_FILE_NAME);
@@ -70,7 +77,7 @@ public class GridFactory {
           }
         }
       }
-      return new Grid(grid);
+      return new Grid(grid, this.outputAdapter);
     } catch (JSONException e) {
       throw new NeighborhoodRuntimeException(ExceptionKeys.INVALID_GRID);
     }
@@ -85,6 +92,6 @@ public class GridFactory {
         grid[i][j] = new GridSquare(1, 0);
       }
     }
-    return new Grid(grid);
+    return new Grid(grid, this.outputAdapter);
   }
 }
