+++ org-code-javabuilder/neighborhood/src/test/java/org/code/neighborhood/DirectionTest.java
@@ -9,19 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 public class DirectionTest {
-  private final PrintStream standardOut = System.out;
-  private final ByteArrayOutputStream outputStreamCaptor = new ByteArrayOutputStream();
-
-  @BeforeEach
-  public void setUp() {
-    System.setOut(new PrintStream(outputStreamCaptor));
-  }
-
-  @AfterEach
-  public void tearDown() {
-    System.setOut(standardOut);
-  }
-
   @Test
   void constructorIgnoresCase() {
     Direction dir = Direction.fromString("NoRtH");
@@ -45,7 +32,6 @@ public class DirectionTest {
     Direction dir = Direction.fromString("North");
     dir = dir.turnLeft();
     assertTrue(dir.isWest());
-    assertTrue(outputStreamCaptor.toString().trim().contains("pointing west"));
   }
 
   @Test
@@ -53,6 +39,5 @@ public class DirectionTest {
     Direction dir = Direction.fromString("East");
     dir = dir.turnLeft();
     assertTrue(dir.isNorth());
-    assertTrue(outputStreamCaptor.toString().trim().contains("pointing north"));
   }
 }
