+++ org-code-javabuilder/neighborhood/src/test/java/org/code/neighborhood/GridFactoryTest.java
@@ -1,8 +1,11 @@
 package org.code.neighborhood;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
 
 import java.io.IOException;
+
+import org.code.protocol.OutputAdapter;
 import org.junit.jupiter.api.Test;
 
 public class GridFactoryTest {
@@ -11,7 +14,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromString() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Grid grid = null;
     try {
       grid = gridFactory.createGridFromString(sampleGrid);
@@ -24,7 +27,7 @@ public class GridFactoryTest {
 
   @Test
   void createEmptyGrid() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Grid grid = gridFactory.createEmptyGrid(2);
     assertTrue(grid instanceof Grid);
     assertTrue(grid.validLocation(1, 1));
@@ -32,7 +35,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromStringWithInvalidJSONThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -46,7 +49,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromStringWithInvalidGridShapeThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -60,7 +63,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromNotSquareGridThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -74,7 +77,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromStringWithInvalidTileTypeThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -87,7 +90,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromStringWithInvalidAssetIdThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -100,7 +103,7 @@ public class GridFactoryTest {
 
   @Test
   void createGridFromStringWithInvalidValueThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
@@ -114,7 +117,7 @@ public class GridFactoryTest {
 
   @Test
   void creatingEmptyGridThrowsException() {
-    GridFactory gridFactory = new GridFactory();
+    GridFactory gridFactory = new GridFactory(mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
