+++ org-code-javabuilder/neighborhood/src/test/java/org/code/neighborhood/GridTest.java
@@ -1,7 +1,9 @@
 package org.code.neighborhood;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
 
+import org.code.protocol.OutputAdapter;
 import org.junit.jupiter.api.Test;
 
 public class GridTest {
@@ -11,7 +13,7 @@ public class GridTest {
     GridSquare s = new GridSquare(1, 0);
     GridSquare[][] squares = new GridSquare[1][1];
     squares[0][0] = s;
-    Grid grid = new Grid(squares);
+    Grid grid = new Grid(squares, mock(OutputAdapter.class));
   }
 
   @Test
@@ -19,7 +21,7 @@ public class GridTest {
     GridSquare s = new GridSquare(1, 0);
     GridSquare[][] squares = new GridSquare[1][1];
     squares[0][0] = s;
-    Grid grid = new Grid(squares);
+    Grid grid = new Grid(squares, mock(OutputAdapter.class));
     assertTrue(grid.validLocation(0, 0));
     assertFalse(grid.validLocation(0, 1));
     assertFalse(grid.validLocation(1, 0));
@@ -32,7 +34,7 @@ public class GridTest {
     GridSquare s = new GridSquare(1, 0);
     GridSquare[][] squares = new GridSquare[1][1];
     squares[0][0] = s;
-    Grid grid = new Grid(squares);
+    Grid grid = new Grid(squares, mock(OutputAdapter.class));
     GridSquare sq = grid.getSquare(0, 0);
     assertEquals(sq, s);
   }
@@ -42,7 +44,7 @@ public class GridTest {
     GridSquare s = new GridSquare(1, 0);
     GridSquare[][] squares = new GridSquare[1][1];
     squares[0][0] = s;
-    Grid grid = new Grid(squares);
+    Grid grid = new Grid(squares, mock(OutputAdapter.class));
     Exception exception =
         assertThrows(
             NeighborhoodRuntimeException.class,
